feat: enable Firestore persistence
